+++ libs/mops/src/thds/mops/pure/__init__.py
@@ -5,6 +5,7 @@
 # to avoid requiring the additional dependency.
 
 from . import adls  # noqa
+from ._futures import MopsFuture  # noqa
 from ._magic.api import magic  # noqa
 from ._magic.sauce import Wand  # noqa
 from .core.entry import register_entry_handler